Move pytest fixtures out of conftest.py and into separate plugin packages.
